added exception handler and rbac

+++ PlaceMicroservice/src/main/java/com/place/PlaceMicroservice/controller/PlaceController.java
@@ -2,6 +2,7 @@ package com.place.PlaceMicroservice.controller;
 
 import com.place.PlaceMicroservice.entity.Place;
 import com.place.PlaceMicroservice.service.PlaceService;
+import com.place.PlaceMicroservice.util.AccessControlUtil;
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import io.github.resilience4j.ratelimiter.annotation.RateLimiter;
 import io.github.resilience4j.retry.annotation.Retry;
@@ -20,8 +21,11 @@ public class PlaceController {
     private PlaceService placeService;
 
     @PostMapping
-    public ResponseEntity<Place> addPlace(@RequestBody Place place){
-        return ResponseEntity.status(HttpStatus.CREATED).body(placeService.addPlace(place));
+    public ResponseEntity<Place> addPlace( @RequestBody Place place,
+                                           @RequestHeader("X-User-Id") String requesterId
+    ) {
+        Place saved = placeService.addPlace(place, requesterId);
+        return ResponseEntity.status(HttpStatus.CREATED).body(saved);
     }
 
 
@@ -68,26 +72,51 @@ public class PlaceController {
     }
 
     @PutMapping(value="/id/{placeId}")
-    public ResponseEntity<Place> updatePlaceByPlaceId(@PathVariable String placeId, @RequestBody Place place) {
+    public ResponseEntity<Place> updatePlaceByPlaceId(@PathVariable String placeId, @RequestBody Place place,
+                                                      @RequestHeader("X-User-Id") String requesterId,
+                                                      @RequestHeader("X-User-Role") String role) {
+        Place existingPlace = placeService.getPlaceByPlaceId(placeId);
+        if (!AccessControlUtil.isAdminOrSelf(role, requesterId, existingPlace.getCreatedBy())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
         Place updatedPlace = placeService.updatePlaceByPlaceId(placeId, place);
         return ResponseEntity.ok(updatedPlace);
     }
 
     @PutMapping(value="/name/{placeName}")
-    public ResponseEntity<Place> updatePlaceByPlaceName(@PathVariable String placeName, @RequestBody Place place) {
+    public ResponseEntity<Place> updatePlaceByPlaceName(@PathVariable String placeName, @RequestBody Place place,
+                                                        @RequestHeader("X-User-Id") String requesterId,
+                                                        @RequestHeader("X-User-Role") String role) {
+        Place existingPlace = placeService.getPlaceByPlaceId(placeName);
+        if (!AccessControlUtil.isAdminOrSelf(role, requesterId, existingPlace.getCreatedBy())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
         Place updatedPlace = placeService.updatePlaceByPlaceName(placeName, place);
         return ResponseEntity.ok(updatedPlace);
     }
 
 
     @DeleteMapping("/id/{placeID}")
-    public ResponseEntity<String> deletePlaceByPlaceId(@PathVariable String placeID) {
-        placeService.deletePlaceByPlaceId(placeID);
+    public ResponseEntity<String> deletePlaceByPlaceId(@PathVariable String placeId,
+                                                       @RequestHeader("X-User-Id") String requesterId,
+                                                       @RequestHeader("X-User-Role") String role) {
+        Place place = placeService.getPlaceByPlaceId(placeId);
+
+        if (!AccessControlUtil.isAdminOrSelf(role, requesterId, place.getCreatedBy())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied: Only creator or admin can delete this place.");
+        }
+        placeService.deletePlaceByPlaceId(placeId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Place deleted successfully");
     }
 
     @DeleteMapping("/name/{placeName}")
-    public ResponseEntity<String> deletePlaceByPlaceName(@PathVariable String placeName) {
+    public ResponseEntity<String> deletePlaceByPlaceName(@PathVariable String placeName,
+                                                         @RequestHeader("X-User-Id") String requesterId,
+                                                         @RequestHeader("X-User-Role") String role) {
+        Place place = placeService.getPlaceByPlaceName(placeName);
+        if (!AccessControlUtil.isAdminOrSelf(role, requesterId, place.getCreatedBy())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied: Only creator or admin can delete this place.");
+        }
         placeService.deletePlaceByPlaceName(placeName);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Place deleted successfully");
     }
